pyproject.toml
[project]
name = "notebooklm-john"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "beautifulsoup4>=4.14.2",
    "dotenv>=0.9.9",
    "fake-useragent>=2.2.0",
    "google-auth>=2.40.3",
    "google-genai>=1.33.0",
    "lxml>=6.0.1",
    "numpy>=2.3.4",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "playwright>=1.55.0",
    "playwright-stealth>=2.0.0",
    "psutil>=7.1.1",
    "pymupdf>=1.26.5",
    "pypdf2>=3.0.1",
    "python-pptx>=1.0.2",
]
